SonarQube batch 6: JS modernization + C++ fixes (~480 issues)
JavaScript rules fixed:
- S7773 (221): parseInt/parseFloat -> Number.parseInt/Number.parseFloat
- S7748 (101): Remove zero fractions (1.0 -> 1, 5.0 -> 5)
- S7761 (44): getAttribute('data-x') -> .dataset.x with camelCase
- S6582 (22): Optional chaining (x && x.y -> x?.y)
- S7781 (14): .replace(/x/g) -> .replaceAll()
- S7764 (22): window -> globalThis
- S7735 (15): Negated conditions -> positive conditions
- S6660 (2): if-only-in-else -> else-if
- S1126 (1): if-then-else -> single return
- S1481+S1854 (12): Remove unused vars and dead stores
- S7762 (4): parentNode.removeChild -> .remove()
- S4043 (2): Array sort to separate statement
- S7756 (2): FileReader -> Blob.text()
- S7770 (1): Arrow fn -> Boolean
- S2871 (1): Sort with localeCompare
- S7778 (1): Merge push calls
- S7723 (1): Array() -> new Array()

C++ rules fixed:
- S1103 (7): Remove misleading /* in comments
- S125 (2): Remove commented-out code references
- S5276 (5): Explicit casts for narrowing conversions
- S3574 (1): Remove redundant lambda return type
- S5350 (1): const pointer in exportToJson loop

Build: SUCCESS | Tests: 180/180 PASSED

+++ src/communication/CommandDispatcher.cpp
@@ -389,9 +389,9 @@ bool CommandDispatcher::handleDecelZoneCommands(const char* cmd, JsonDocument& d
         if (doc["randomTurnbackEnabled"].is<bool>()) {
             zoneEffect.randomTurnbackEnabled = doc["randomTurnbackEnabled"].as<bool>();
         }
-        int turnbackChance = doc["turnbackChance"] | zoneEffect.turnbackChance;
+        int turnbackChance = doc["turnbackChance"] | static_cast<int>(zoneEffect.turnbackChance);
         if (turnbackChance >= 0 && turnbackChance <= 100) {
-            zoneEffect.turnbackChance = turnbackChance;
+            zoneEffect.turnbackChance = static_cast<uint8_t>(turnbackChance);
         }
 
         // === End Pause ===
